Refactor structure of files. Extract some code to separate view classes
