+++ kotlin-eclipse-ui/plugin.xml
@@ -218,7 +218,7 @@
       </page>
       <page
             category="kotlin-eclipse-ui.preferences"
-            class="org.jetbrains.kotlin.preferences.GlobalCompilerPropertyPage"
+            class="org.jetbrains.kotlin.preferences.WorkspaceCompilerPropertyPage"
             id="kotlin-eclipse-ui.preferences.compiler"
             name="Compiler">
       </page>
@@ -892,7 +892,7 @@
    <extension
          point="org.eclipse.ui.propertyPages">
       <page
-            class="org.jetbrains.kotlin.preferences.properties.ProjectCompilerPropertyPage"
+            class="org.jetbrains.kotlin.preferences.ProjectCompilerPropertyPage"
             id="org.jetbrains.kotlin.ui.project-properties"
             name="Kotlin Compiler">
          <enabledWhen>
