+++ kotlin-eclipse-ui/src/org/jetbrains/kotlin/preferences/views/Validable.kt
@@ -1,0 +1,6 @@
+package org.jetbrains.kotlin.preferences.views
+
+interface Validable {
+    val isValid: Boolean
+    var onIsValidChanged: (Boolean) -> Unit
+}
